fix: resolve container listing authentication and exception issues
- Fix API key authentication in MCP server HTTP client calls
- Fix ContainerError parameter mismatch (device_id -> hostname)
- Remove deprecated monitoring.py endpoint
- Add missing system monitoring endpoints to devices router
- All MCP tools now working correctly with 110+ containers
